securedrop-admin: implement dev-shell
The dev-shell script refresh the securedrop-admin docker image if the
requirements or the bootstrap script are updated. It is done by
running docker build at the begining of the script and the
dependencies mechanism do the rest. The resulting image has an up to
date virtualenv in /opt/.venv ready to use and all the system
dependencies installed.

docker run has the admin directory mounted from the host into the
container, at the same location to avoid any trouble with hardcoded
absolute paths. It runs with the id of the host user so files created
in the container are owned by the host user instead of root. If a
script needs root access, it has passwordless sudo permissions.

Convenience makefile targets (make test, ...) are converted to use
dev-shell.

Example usages:

Run only flake8 with

     bin/dev-shell tox -e flake8

Run only one test foobar with

     bin/dev-shell tox -e py2 -- -k foobar